+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/KafkaBridgeCluster.java
@@ -86,12 +86,12 @@ public class KafkaBridgeCluster extends AbstractModel implements SupportsLogging
      * HTTP port configuration
      */
     public static final int DEFAULT_REST_API_PORT = 8080;
-
     /* test */ static final String COMPONENT_TYPE = "kafka-bridge";
     protected static final String REST_API_PORT_NAME = "rest-api";
     /* test */ static final String REST_API_MANAGEMENT_PORT_NAME = "rest-api-mgmt";
     /* test */ static final int REST_API_MANAGEMENT_PORT = 8081;
     protected static final String TLS_CERTS_BASE_VOLUME_MOUNT = "/opt/strimzi/bridge-certs/";
+    protected static final String HTTP_SERVER_CERTS_BASE_VOLUME_MOUNT = "/opt/strimzi/bridge-http-certs/";
     protected static final String PASSWORD_VOLUME_MOUNT = "/opt/strimzi/bridge-password/";
     protected static final String ENV_VAR_KAFKA_INIT_INIT_FOLDER_KEY = "INIT_FOLDER";
     private static final String KAFKA_BRIDGE_CONFIG_VOLUME_NAME = "kafka-bridge-configurations";
@@ -246,8 +246,10 @@ public class KafkaBridgeCluster extends AbstractModel implements SupportsLogging
      */
     public Service generateService() {
         int port = DEFAULT_REST_API_PORT;
+        boolean isTls = false;
         if (http != null) {
             port = http.getPort();
+            isTls = http.getTls() != null;
         }
 
         List<ServicePort> ports = new ArrayList<>();
@@ -263,7 +265,7 @@ public class KafkaBridgeCluster extends AbstractModel implements SupportsLogging
                 ports,
                 labels.strimziSelectorLabels(),
                 ModelUtils.getCustomLabelsOrAnnotations(CO_ENV_VAR_CUSTOM_SERVICE_LABELS),
-                Util.mergeLabelsOrAnnotations(getDiscoveryAnnotation(port), ModelUtils.getCustomLabelsOrAnnotations(CO_ENV_VAR_CUSTOM_SERVICE_ANNOTATIONS))
+                Util.mergeLabelsOrAnnotations(getDiscoveryAnnotation(port, isTls), ModelUtils.getCustomLabelsOrAnnotations(CO_ENV_VAR_CUSTOM_SERVICE_ANNOTATIONS))
         );
     }
 
@@ -272,15 +274,14 @@ public class KafkaBridgeCluster extends AbstractModel implements SupportsLogging
      *
      * @return  JSON with discovery annotation
      */
-    /*test*/ Map<String, String> getDiscoveryAnnotation(int port) {
+    /*test*/ Map<String, String> getDiscoveryAnnotation(int port, boolean isTls) {
         JsonArray anno = new JsonArray();
 
         JsonObject discovery = new JsonObject();
         discovery.put("port", port);
-        discovery.put("tls", false);
+        discovery.put("tls", isTls);
         discovery.put("auth", "none");
-        discovery.put("protocol", "http");
-
+        discovery.put("protocol", isTls ? "https" : "http");
         anno.add(discovery);
 
         JsonObject managementDiscovery = new JsonObject();
@@ -316,6 +317,19 @@ public class KafkaBridgeCluster extends AbstractModel implements SupportsLogging
             CertUtils.createTrustedCertificatesVolumes(volumeList, tls.getTrustedCertificates(), isOpenShift);
         }
 
+        if (http != null && http.getTls() != null) {
+            if (http.getTls().getCertificateAndKey() != null) {
+                String secretName = http.getTls().getCertificateAndKey().getSecretName();
+                // skipping if a volume mount with same Secret name was already added
+                if (volumeList.stream().noneMatch(v -> v.getName().equals(secretName))) {
+                    volumeList.add(VolumeUtils.createSecretVolume(secretName, secretName, isOpenShift));
+                }
+            } else {
+                LOGGER.warnCr(reconciliation, "The spec.http.tls.certificateAndKey property is missing. This is required when TLS is enabled for HTTP Bridge.");
+                throw new InvalidResourceException("The TLS configuration for the HTTP is not specified.");
+            }
+        }
+
         if (rack != null) {
             volumeList.add(VolumeUtils.createEmptyDirVolume(INIT_VOLUME_NAME, "1Mi", "Memory"));
         }
@@ -335,6 +349,16 @@ public class KafkaBridgeCluster extends AbstractModel implements SupportsLogging
 
         if (tls != null) {
             CertUtils.createTrustedCertificatesVolumeMounts(volumeMountList, tls.getTrustedCertificates(), TLS_CERTS_BASE_VOLUME_MOUNT);
+        }
+
+        if (http != null && http.getTls() != null) {
+            if (http.getTls().getCertificateAndKey() != null) {
+                String secretName = http.getTls().getCertificateAndKey().getSecretName();
+                volumeMountList.add(VolumeUtils.createVolumeMount(secretName, HTTP_SERVER_CERTS_BASE_VOLUME_MOUNT + secretName));
+            } else {
+                LOGGER.warnCr(reconciliation, "The spec.http.tls.certificateAndKey property is missing. This is required when TLS is enabled for HTTP Bridge.");
+                throw new InvalidResourceException("The TLS configuration for the HTTP is not specified.");
+            }
         }
 
         if (rack != null) {