+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/KafkaBridgeClusterTest.java
@@ -214,7 +214,7 @@ public class KafkaBridgeClusterTest {
         assertThat(svc.getSpec().getIpFamilyPolicy(), is(nullValue()));
         assertThat(svc.getSpec().getIpFamilies(), is(nullValue()));
 
-        assertThat(svc.getMetadata().getAnnotations(), is(KBC.getDiscoveryAnnotation(KafkaBridgeCluster.DEFAULT_REST_API_PORT)));
+        assertThat(svc.getMetadata().getAnnotations(), is(KBC.getDiscoveryAnnotation(KafkaBridgeCluster.DEFAULT_REST_API_PORT, false)));
 
         io.strimzi.operator.cluster.TestUtils.checkOwnerReference(svc, RESOURCE);
     }
@@ -420,6 +420,40 @@ public class KafkaBridgeClusterTest {
         String bridgeConfigurations = configMap.getData().get(BRIDGE_CONFIGURATION_FILENAME);
         assertThat(bridgeConfigurations, containsString("kafka.sasl.mechanism=PLAIN"));
         assertThat(bridgeConfigurations, containsString("kafka.sasl.jaas.config=org.apache.kafka.common.security.plain.PlainLoginModule required username=\"user1\" password=\"${strimzidir:/opt/strimzi/bridge-password/user1-secret:password}\";"));
+    }
+
+    @Test
+    public void testGenerateDeploymentWithHttpTls() {
+        KafkaBridge resource = new KafkaBridgeBuilder(RESOURCE)
+                .editSpec()
+                    .withNewHttp()
+                        .withPort(8443)
+                        .withNewTls()
+                            .withNewCertificateAndKey()
+                                .withSecretName("my-secret")
+                                .withCertificate("my-secret")
+                                .withKey("private.key")
+                            .endCertificateAndKey()
+                        .endTls()
+                    .endHttp()
+                .endSpec()
+                .build();
+        KafkaBridgeCluster kbc = KafkaBridgeCluster.fromCrd(Reconciliation.DUMMY_RECONCILIATION, resource, SHARED_ENV_PROVIDER);
+        Deployment dep = kbc.generateDeployment(emptyMap(), true, null, null);
+
+        assertThat(dep.getSpec().getTemplate().getSpec().getVolumes().get(2).getName(), is("my-secret"));
+
+        List<Container> containers = dep.getSpec().getTemplate().getSpec().getContainers();
+
+        assertThat(containers.get(0).getVolumeMounts().get(2).getMountPath(), is(KafkaBridgeCluster.HTTP_SERVER_CERTS_BASE_VOLUME_MOUNT + "my-secret"));
+
+        ConfigMap configMap = kbc.generateBridgeConfigMap(METRICS_AND_LOGGING);
+        String bridgeConfigurations = configMap.getData().get(BRIDGE_CONFIGURATION_FILENAME);
+
+        assertThat(bridgeConfigurations, containsString("http.port=8443"));
+        assertThat(bridgeConfigurations, containsString("http.ssl.enable=true"));
+        assertThat(bridgeConfigurations, containsString("http.ssl.certificate.location=/opt/strimzi/bridge-http-certs/my-secret/my-secret"));
+        assertThat(bridgeConfigurations, containsString("http.ssl.key.location=/opt/strimzi/bridge-http-certs/my-secret/private.key"));
     }
 
     @Test
@@ -1039,7 +1073,7 @@ public class KafkaBridgeClusterTest {
         assertThat(svc.getSpec().getPorts().get(0).getPort(), is(1874));
         assertThat(svc.getSpec().getPorts().get(0).getName(), is(KafkaBridgeCluster.REST_API_PORT_NAME));
         assertThat(svc.getSpec().getPorts().get(0).getProtocol(), is("TCP"));
-        assertThat(svc.getMetadata().getAnnotations(), is(KBC.getDiscoveryAnnotation(1874)));
+        assertThat(svc.getMetadata().getAnnotations(), is(KBC.getDiscoveryAnnotation(1874, false)));
         assertThat(svc.getSpec().getPorts().get(1).getPort(), is(KafkaBridgeCluster.REST_API_MANAGEMENT_PORT));
         assertThat(svc.getSpec().getPorts().get(1).getName(), is(KafkaBridgeCluster.REST_API_MANAGEMENT_PORT_NAME));
         io.strimzi.operator.cluster.TestUtils.checkOwnerReference(svc, resource);